+++ docs/history.md
@@ -13,6 +13,7 @@ N/A
 * `accounts-2fa@1.0.1`
   - Reduce one DB call on 2FA login. [PR](https://github.com/meteor/meteor/pull/11985)
   - Throw error when user is not found on `Accounts._is2faEnabledForUser`
+  - Remove vulnerability from the method `has2faEnabled`
 * `accounts-password@2.3.1`
   - Use method `Accounts._check2faEnabled` when validating 2FA
 * `accounts-passwordless@2.1.1`