+++ docs/history.md
@@ -4,7 +4,7 @@
 
 #### Breaking Changes
 
-* `accounts-2fa@1.0.1`
+* `accounts-2fa@2.0.0`
   - The method `has2faEnabled` no longer takes a selector as an argument, just the callback.
   - `generate2faActivationQrCode` now throws an error if it's being called when the user already has 2FA enabled.
 
@@ -12,7 +12,7 @@
 
 #### Meteor Version Release
 
-* `accounts-2fa@1.0.1`
+* `accounts-2fa@2.0.0`
   - Reduce one DB call on 2FA login. [PR](https://github.com/meteor/meteor/pull/11985)
   - Throw error when user is not found on `Accounts._is2faEnabledForUser`
   - Remove vulnerability from the method `has2faEnabled`